Added nodemon, Page routes were created,
We began our blog project with an existing blog template from
startbootstrap.com and incorporated it into a Node.js project. Using the nodemon package, we automatically detect code changes in our project and restart the server. Page routes were created to serve the home, about and postpages. A navigation bar was created to contain the links to these routes.