Added Form, body-parser, displayed the list of blog posts in the home page
We used MongoDB to build our blog app. We implemented a form to create a blog post and used the body-parser package to retrieve the form field data. The BlogPost model was used to store the data in the database. We displayedthe list of blog posts in the home page with the EJS templating engine and also displayed each blog post detail in their own individual pages.